+++ src/components/ReportModals/CashDrawerReportModal.test.js
@@ -1,14 +1,15 @@
 import React from 'react';
 import { Router } from 'react-router-dom';
 import {
-  render,
   cleanup,
+  render,
   screen,
 } from '@testing-library/react';
 import { createMemoryHistory } from 'history';
+import { forEach } from 'lodash';
 
 import '../../../test/jest/__mock__';
-import CashDrawerReportModal from './CashDrawerReportModal';
+import CashDrawerReportModal, { validate } from './CashDrawerReportModal';
 
 const modalHeader = 'Cash drawer reconciliation modal';
 const renderCashDrawerReportModal = ({
@@ -17,7 +18,7 @@ const renderCashDrawerReportModal = ({
   handleSubmit = jest.fn(),
   values = {},
   cashDrawerReportSources = {
-    POST: jest.fn()
+    POST: jest.fn(() => Promise.resolve({ data: {} })),
   },
   intl = {},
   form = {},
@@ -43,6 +44,7 @@ const renderCashDrawerReportModal = ({
         intl={intl}
         form={form}
         cashDrawerReportSources={cashDrawerReportSources}
+        dismissible="true"
       />
     </Router>
   );
@@ -56,7 +58,6 @@ describe('Cash drawer reconciliation modal', () => {
       cashDrawerModal = renderCashDrawerReportModal({
         initialValues: {
           format: 'both',
-          servicePoint: 'Online',
           sources: ['ADMIN'],
           startDate: '2020-05-11',
           endDate: '2021-05-11'
@@ -165,6 +166,83 @@ describe('Cash drawer reconciliation modal', () => {
 
       expect(container).toBeVisible();
       expect(form).toBeVisible();
+    });
+  });
+
+  describe('Validation', () => {
+    const customRender = (options) => {
+      let validationError = '';
+
+      forEach(validate(options), (error) => {
+        validationError = error;
+      });
+
+      return render(
+        <div data-testid="validation-error">
+          {
+            validationError
+          }
+        </div>,
+      );
+    };
+
+    it('should return validation error for start date', () => {
+      const options = {
+        startDate: '',
+        endDate: '',
+        servicePoint: [],
+      };
+
+      customRender(options);
+
+      expect(screen.getByTestId('validation-error').textContent).toBe('ui-users.reports.cash.drawer.report.startDate.error');
+    });
+
+    it('should return validation error for end date without start date', () => {
+      const options = {
+        startDate: '',
+        endDate: '2021-06-15',
+        servicePoint: [],
+      };
+
+      customRender(options);
+
+      expect(screen.getByTestId('validation-error').textContent).toBe('ui-users.reports.cash.drawer.report.endDateWithoutStart.error');
+    });
+
+    it('should return validation error for end date', () => {
+      const options = {
+        startDate: '2021-06-30',
+        endDate: '2021-06-15',
+        servicePoint: [],
+      };
+
+      customRender(options);
+
+      expect(screen.getByTestId('validation-error').textContent).toBe('ui-users.reports.cash.drawer.report.endDate.error');
+    });
+
+    it('should return validation error for service point', () => {
+      const options = {
+        startDate: '2021-06-15',
+        endDate: '2021-06-30',
+      };
+
+      customRender(options);
+
+      expect(screen.getByTestId('validation-error').textContent).toBe('ui-users.reports.cash.drawer.report.servicePoint.error');
+    });
+
+    it('should not return validation error', () => {
+      const options = {
+        startDate: '2021-06-15',
+        endDate: '2021-06-30',
+        servicePoint: [],
+      };
+
+      customRender(options);
+
+      expect(screen.getByTestId('validation-error').textContent).toBe('');
     });
   });
 });