+++ src/views/UserDetail/UserDetail.js
@@ -11,6 +11,7 @@ import { injectIntl, FormattedMessage } from 'react-intl';
 
 import {
   AppIcon,
+  CalloutContext,
   IfPermission,
   IfInterface,
   TitleManager,
@@ -34,6 +35,7 @@ import {
 import {
   NotesSmartAccordion,
   ViewCustomFieldsRecord,
+  NotePopupModal,
 } from '@folio/stripes/smart-components';
 
 import {
@@ -62,6 +64,8 @@ import {
 import RequestFeeFineBlockButtons from '../../components/RequestFeeFineBlockButtons';
 import { departmentsShape } from '../../shapes';
 
+import OpenTransactionModal from './components/OpenTransactionModal';
+import DeleteUserModal from './components/DeleteUserModal';
 import ExportFeesFinesReportButton from './components';
 import ErrorPane from '../../components/ErrorPane';
 
@@ -77,6 +81,8 @@ class UserDetail extends React.Component {
     }).isRequired,
     resources: PropTypes.shape({
       selUser: PropTypes.object,
+      delUser: PropTypes.object,
+      openTransactions: PropTypes.object,
       user: PropTypes.arrayOf(PropTypes.object),
       accounts: PropTypes.shape({
         records: PropTypes.arrayOf(PropTypes.object),
@@ -123,11 +129,14 @@ class UserDetail extends React.Component {
     tagsEnabled: PropTypes.bool,
     paneWidth: PropTypes.string,
     intl: PropTypes.object.isRequired,
+    mutator: PropTypes.object,
   };
 
   static defaultProps = {
     paneWidth: '44%'
   };
+
+  static contextType = CalloutContext;
 
   constructor(props) {
     super(props);
@@ -151,6 +160,8 @@ class UserDetail extends React.Component {
 
     this.state = {
       lastUpdate: null,
+      showOpenTransactionModal: false,
+      showDeleteUserModal: false,
       sections: {
         userInformationSection: true,
         extendedInfoSection: false,
@@ -178,6 +189,34 @@ class UserDetail extends React.Component {
     // Logging below shows this DOES sometimes find the wrong record. But why?
     // console.log(`getUser: found ${selUser.length} users, id '${selUser[0].id}' ${selUser[0].id === id ? '==' : '!='} '${id}'`);
     return selUser.find(u => u.id === id);
+  }
+
+  handleDeleteUser = (id) => {
+    const { history, location, mutator, intl } = this.props;
+    const user = this.getUser();
+    const fullNameOfUser = getFullName(user);
+
+    try {
+      mutator.delUser.DELETE({ id })
+        .then(() => history.replace(
+          {
+            pathname: '/users',
+            search: `${location.search}`,
+            state: { deletedUserId: id }
+          }
+        ))
+        .then(() => {
+          this.context.sendCallout({
+            type: 'success',
+            message: intl.formatMessage({ id: 'ui-users.details.deleteUser.success' }, { name: fullNameOfUser }),
+          });
+        });
+    } catch (error) {
+      this.context.sendCallout({
+        type: 'error',
+        message: intl.formatMessage({ id: 'ui-users.details.deleteUser.failed' }, { name: fullNameOfUser }),
+      });
+    }
   }
 
   checkScope = () => {
@@ -308,6 +347,45 @@ class UserDetail extends React.Component {
     );
   }
 
+  showOpenTransactionsModal(json) {
+    this.setState({
+      showOpenTransactionModal: true,
+      openTransactions: json,
+    });
+  }
+
+  showDeleteUserModal(json) {
+    this.setState({
+      showDeleteUserModal: true,
+      openTransactions: json,
+    });
+  }
+
+  doCloseTransactionDeleteModal = () => {
+    this.setState({
+      showDeleteUserModal: false,
+      showOpenTransactionModal: false,
+    });
+  }
+
+  selectModal(transactions) {
+    if (!transactions.hasOpenTransactions) {
+      this.showDeleteUserModal();
+    } else {
+      this.showOpenTransactionsModal(transactions);
+    }
+  }
+
+  handleDeleteClick() {
+    const { mutator } = this.props;
+    const userId = this.props.match.params.id;
+
+    mutator.openTransactions.GET({ userId })
+      .then((response) => {
+        this.selectModal(response);
+      });
+  }
+
   getActionMenu = barcode => ({ onToggle }) => {
     const {
       okapi: {
@@ -335,7 +413,8 @@ class UserDetail extends React.Component {
     const showActionMenu = this.props.stripes.hasPerm('ui-users.edit')
       || this.props.stripes.hasPerm('ui-users.patron_blocks')
       || this.props.stripes.hasPerm('ui-users.feesfines.actions.all')
-      || this.props.stripes.hasPerm('ui-requests.all');
+      || this.props.stripes.hasPerm('ui-requests.all')
+      || this.props.stripes.hasPerm('ui-users.delete,ui-users.opentransactions');
 
     if (showActionMenu) {
       return (
@@ -370,6 +449,21 @@ class UserDetail extends React.Component {
               callout={this.callout}
             />
           </IfInterface>
+          <IfPermission perm="ui-users.delete,ui-users.opentransactions">
+            <Button
+              buttonStyle="dropdownItem"
+              data-test-actions-menu-check-delete
+              id="clickable-checkdeleteuser"
+              onClick={() => {
+                this.handleDeleteClick();
+                onToggle();
+              }}
+            >
+              <Icon icon="trash">
+                <FormattedMessage id="ui-users.details.checkDelete" />
+              </Icon>
+            </Button>
+          </IfPermission>
         </>
       );
     } else {
@@ -430,6 +524,8 @@ class UserDetail extends React.Component {
     } = this.state;
 
     const user = this.getUser();
+    const fullNameOfUser = getFullName(user);
+    const userId = match.params.id;
 
     const addressTypes = (resources.addressTypes || {}).records || [];
     const addresses = getFormAddressList(get(user, 'personal.addresses', []));
@@ -605,6 +701,7 @@ class UserDetail extends React.Component {
                       location={location}
                       accounts={accounts}
                       match={match}
+                      {...this.props}
                     />
                   </IfPermission>
                 </IfInterface>
@@ -686,6 +783,28 @@ class UserDetail extends React.Component {
             </Pane>
             { helperApp && <HelperApp appName={helperApp} onClose={this.closeHelperApp} /> }
             <Callout ref={(ref) => { this.callout = ref; }} />
+            <NotePopupModal
+              id="user-popup-note-modal"
+              domainName="users"
+              entityType="user"
+              popUpPropertyName="popUpOnUser"
+              entityId={user?.id}
+            />
+            {this.state.showDeleteUserModal &&
+            <DeleteUserModal
+              onCloseModal={this.doCloseTransactionDeleteModal}
+              username={fullNameOfUser}
+              userId={userId}
+              deleteUser={this.handleDeleteUser}
+            />
+            }
+            {this.state.showOpenTransactionModal &&
+            <OpenTransactionModal
+              onCloseModal={this.doCloseTransactionDeleteModal}
+              openTransactions={this.state.openTransactions}
+              username={fullNameOfUser}
+            />
+            }
           </>
         </HasCommand>
       );