+++ test/jest/__mock__/index.js
@@ -1,5 +1,8 @@
 import './stripesConfig.mock';
+import './stripesCore.mock';
 import './stripes.mock';
 import './intl.mock';
 import './stripesIcon.mock';
 import './stripesComponents.mock';
+import './stripesSmartComponent.mock';
+import './stripesUtils.mock';