+++ test/jest/__mock__/stripesComponents.mock.js
@@ -6,20 +6,44 @@ jest.mock('@folio/stripes/components', () => ({
       <span>{props.children}</span>
     </span>
   )),
-  Button: jest.fn(({ children }) => (
-    <button data-test-button type="button">
+  Button: jest.fn(({ children, onClick = jest.fn() }) => (
+    <button data-test-button type="button" onClick={onClick}>
       <span>
         {children}
       </span>
     </button>
   )),
   Col: jest.fn(({ children }) => <div className="col">{ children }</div>),
+  ExpandAllButton: jest.fn(({ children, ...rest }) => (
+    <span {...rest}>{children}</span>
+  )),
+  AccordionSet: jest.fn(({ children, ...rest }) => (
+    <span {...rest}>{children}</span>
+  )),
+  Accordion: jest.fn(({ children, ...rest }) => (
+    <span {...rest}>{children}</span>
+  )),
+  Callout: jest.fn(({ children, ...rest }) => (
+    <span {...rest}>{children}</span>
+  )),
+  HasCommand: jest.fn(({ children, ...rest }) => (
+    <span {...rest}>{children}</span>
+  )),
+  KeyValue: jest.fn(({ label, children, value }) => (
+    <>
+      <span>{label}</span>
+      <span>{value || children}</span>
+    </>
+  )),
+  NoValue: jest.fn(({ ariaLabel }) => (<span>{ariaLabel}</span>)),
+  Loading: () => <div>Loading</div>,
   Datepicker: jest.fn(({ ref, children, ...rest }) => (
     <div ref={ref} {...rest}>
       {children}
       <input type="text" />
     </div>
   )),
+  Headline: jest.fn(({ children }) => <div>{ children }</div>),
   Icon: jest.fn((props) => (props && props.children ? props.children : <span />)),
   IconButton: jest.fn(({
     buttonProps,
@@ -34,15 +58,29 @@ jest.mock('@folio/stripes/components', () => ({
   Label: jest.fn(({ children, ...rest }) => (
     <span {...rest}>{children}</span>
   )),
-  Modal: jest.fn(({ children, label, ...rest }) => (
-    <div
-      dismissible
-      {...rest}
-    >
-      <h1>{label}</h1>
-      {children}
-    </div>
-  )),
+  // oy, dismissible. we need to pull it out of props so it doesn't
+  // get applied to the div as an attribute, which must have a string-value,
+  // which will shame you in the console:
+  //
+  //     Warning: Received `true` for a non-boolean attribute `dismissible`.
+  //     If you want to write it to the DOM, pass a string instead: dismissible="true" or dismissible={value.toString()}.
+  //         in div (created by mockConstructor)
+  //
+  // is there a better way to throw it away? If we don't destructure and
+  // instead access props.label and props.children, then we get a test
+  // failure that the modal isn't visible. oy, dismissible.
+  Modal: jest.fn(({ children, label, dismissible, footer, ...rest }) => {
+    return (
+      <div
+        data-test={dismissible ? '' : ''}
+        {...rest}
+      >
+        <h1>{label}</h1>
+        {children}
+        {footer}
+      </div>
+    );
+  }),
   ModalFooter: jest.fn((props) => (
     <div>{props.children}</div>
   )),
@@ -55,14 +93,46 @@ jest.mock('@folio/stripes/components', () => ({
             key={option.id || `option-${i}`}
           >
             {option.label}
-          </option>))}
+          </option>
+        ))}
       </select>
       {children}
     </div>
   )),
+  NavList: jest.fn(({ children, className, ...rest }) => (
+    <div className={className} {...rest}>{children}</div>
+  )),
+  NavListItem: jest.fn(({ children, className, ...rest }) => (
+    <div className={className} {...rest}>{children}</div>
+  )),
+  NavListSection: jest.fn(({ children, className, ...rest }) => (
+    <div className={className} {...rest}>{children}</div>
+  )),
+  Pane: jest.fn(({ children, className, defaultWidth, paneTitle, firstMenu, lastMenu, actionMenu, ...rest }) => {
+    return (
+      <div className={className} {...rest} style={{ width: defaultWidth }}>
+        <div>
+          {firstMenu ?? null}
+          {paneTitle}
+          {actionMenu ? actionMenu({ onToggle: jest.fn() }) : null}
+          {lastMenu ?? null}
+        </div>
+        {children}
+      </div>
+    );
+  }),
   PaneFooter: jest.fn(({ ref, children, ...rest }) => (
     <div ref={ref} {...rest}>{children}</div>
   )),
+  PaneHeader: jest.fn(({ paneTitle, firstMenu, lastMenu, actionMenu }) => (
+    <div actionMenu={actionMenu}>
+      {firstMenu ?? null}
+      {paneTitle}
+      {actionMenu ? actionMenu({ onToggle: jest.fn() }) : null}
+      {lastMenu ?? null}
+    </div>
+  )),
+  PaneBackLink: jest.fn(() => <span />),
   PaneMenu: jest.fn((props) => <div>{props.children}</div>),
   RadioButton: jest.fn(({ label, name, ...rest }) => (
     <div>